Fixed UT error and move weightPruningConfig
