+++ neural_compressor/config.py
@@ -664,33 +664,35 @@ class QuantizationAwareTrainingConfig(_BaseQuantizationConfig):
 pruners = [Pruner()]
 
 
-class PruningConfig:
-    def __init__(self, pruners=pruners, initial_sparsity=0.0, target_sparsity=0.97,
-                 max_sparsity_ratio_per_layer=0.98, prune_type="basic_magnitude",
-                 start_epoch=0, end_epoch=4, start_step=0, end_step=0, update_frequency=1.0,
-                 update_frequency_on_step=1, not_to_prune_names=[], prune_domain="global",
-                 names=[], exclude_names=[], prune_layer_type=[], sparsity_decay_type="exp",
-                 pattern="tile_pattern_1x1"):
-        self.weight_compression = DotDict({
-            'initial_sparsity': initial_sparsity,
+class WeightPruningConfig:
+    """
+    similiar to torch optimizer's interface
+    """
+
+    def __init__(self, pruning_configs=[{}],  ##empty dict will use global values
+                 target_sparsity=0.9, pruning_type="snip_momentum", pattern="4x1", op_names=[],
+                 excluded_op_names=[],
+                 start_step=0, end_step=0, pruning_scope="global", pruning_frequency=1,
+                 min_sparsity_ratio_per_op=0.0, max_sparsity_ratio_per_op=0.98,
+                 sparsity_decay_type="exp", pruning_op_types=['Conv', 'Linear'],
+                 **kwargs):
+        self.pruning_configs = pruning_configs
+        self._weight_compression = DotDict({
             'target_sparsity': target_sparsity,
-            'max_sparsity_ratio_per_layer': max_sparsity_ratio_per_layer,
-            'prune_type': prune_type,
-            'start_epoch': start_epoch,
-            'end_epoch': end_epoch,
+            'pruning_type': pruning_type,
+            'pattern': pattern,
+            'op_names': op_names,
+            'excluded_op_names': excluded_op_names,  ##global only
             'start_step': start_step,
             'end_step': end_step,
-            'update_frequency': update_frequency,
-            'update_frequency_on_step': update_frequency_on_step,
-            'not_to_prune_names': not_to_prune_names,
-            'prune_domain': prune_domain,
-            'names': names,
-            'exclude_names': exclude_names,
-            'prune_layer_type': prune_layer_type,
+            'pruning_scope': pruning_scope,
+            'pruning_frequency': pruning_frequency,
+            'min_sparsity_ratio_per_op': min_sparsity_ratio_per_op,
+            'max_sparsity_ratio_per_op': max_sparsity_ratio_per_op,
             'sparsity_decay_type': sparsity_decay_type,
-            'pattern': pattern,
-            'pruners': pruners
+            'pruning_op_types': pruning_op_types,
         })
+        self._weight_compression.update(kwargs)
 
     @property
     def weight_compression(self):
