+++ test/pruning/test_pruning.py
@@ -5,7 +5,8 @@ import torchvision
 import torch.nn as nn
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
-from neural_compressor.pruner.pruning import Pruning, WeightPruningConfig
+from neural_compressor.config import WeightPruningConfig
+from neural_compressor.pruner.pruning import Pruning
 
 
 class TestPruning(unittest.TestCase):
