+++ test/pruning/test_pruning_config.py
@@ -6,7 +6,8 @@ import torch.nn as nn
 
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
-from neural_compressor.pruner.pruning import Pruning, WeightPruningConfig
+from neural_compressor.config import WeightPruningConfig
+from neural_compressor.pruner.pruning import Pruning
 
 
 class TestPytorchPruning(unittest.TestCase):
