+++ test/pruning/test_pruning_regs.py
@@ -6,7 +6,8 @@ import torch.nn as nn
 
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
-from neural_compressor.pruner.pruning import Pruning, WeightPruningConfig
+from neural_compressor.config import WeightPruningConfig
+from neural_compressor.pruner.pruning import Pruning
 
 local_regs_config = [
     {
